+++ src/Notifications/User/RegisteredNotification.php
@@ -32,7 +32,7 @@ class RegisteredNotification extends Notification implements ShouldQueue
     public function toMail(User $user): MailMessage
     {
         return (new MailMessage)
-            ->replyTo('email@wavehack.net', 'WaveHack')
+            ->replyTo('internetfett@gmail.com', 'InternetFett')
             ->subject('OpenDominion Registration')
             ->greeting('OpenDominion Registration')
             ->line('Hello ' . $user->display_name . '!')